1213 - Ability to add captions to video nodes (#1215)
- Adds captions to the UrlVideoMedia player.
- When a video is uploaded to Kaltura, its captions will also be
uploaded to Kaltura and associated with the uploaded video. The captions
will be deleted locally after successful upload.
- Changes how users create Kaltura videos: to change the Kaltura ID, the
user must submit it in the NodeModal, which loads data from Kaltura into
the NodeModal to be edited.
- Depends on the package
[iso-639-1](https://www.npmjs.com/package/iso-639-1) to get ISO 639-1
language names and language codes and convert between the two.

**Changes to the video authoring form:**
- For mp4 and Kaltura videos, show a toggle for captions (on/off) that
is off by default
- If the toggle is on, displays one caption row by default, with the
option to add more.
- Each caption contains:
- A file upload field for the caption source. Must be VTT for regular
videos, and VTT or SRT for Kaltura videos.
    - A language field. English by default.
- An optional field for the caption label, only visible if the user
turns on "Customize label". Otherwise, the name of the language is used
as the caption label (e.g. "English", "French").
- A "Display in player" toggle (true by default). The caption should
only be displayed in the video player if this is checked.
    - A button to download the caption.
- A video can have a default caption. The “Set as default” button next
to each caption sets the default caption.

When a Kaltura video node is submitted, the new list of captions
replaces the existing list of captions on Kaltura.

**Pending captions:**
- If an error occurs when uploading a caption to Kaltura (either an
error with the .vtt file or the caption metadata), the caption will be
saved as a "pending caption" and the user will receive the number of
captions that failed to upload. The user can edit pending captions in
the NodeModal and add them back to the list. Pending captions are
identified by ID, so that adding a pending caption back to the list
overwrites the caption with the same ID, if any.

### Details
- The set of languages supported by Kaltura is different from the set of
languages in ISO 639-1 (see here
https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes). We offer a union
of Kaltura's language names and the ISO 639-1 language names to users.
If the video is a URL video but the caption language does not exist in
ISO 639-1, the caption's `srclang` attribute will be "". If the video is
a Kaltura video but the caption language does not exist in Kaltura, the
label will be set to the language name (unless the user gave a custom
label) while the language will be "Undefined".
- Since captions do not change often, and to make captions accessible
when Kaltura is not available in the backend, captions uploaded to
Kaltura are also saved under the node's `typeData.captions`. To keep
this information up to date, the captions are fetched from Kaltura every
time the node is opened for editing.

## Screenshot
New editing workflow for Kaltura videos, and editing captions:
https://user-images.githubusercontent.com/44929104/187580872-643b3996-7b1d-4870-b749-f165174fb940.mp4

Caption upload feedback on Kaltura upload dashboard:
![image](https://user-images.githubusercontent.com/44929104/182502434-50472478-e97e-4790-96e2-1e496303e709.png)

Pending captions:
![image](https://user-images.githubusercontent.com/44929104/185474436-4824b7c3-eea6-4aa7-95fe-88bd7a543de4.png)

---------

Co-authored-by: Legendword <zwh@legendword.com>
Co-authored-by: Aidin Niavarani <aidin@wynnset.com>
Co-authored-by: langara2020sz <szhang029@mylangara.ca>